hwserver/requirements.txt
spidev==3.2
nanomsg==1.0
pyserial==3.3
gevent==1.2.2
RPi.GPIO==0.6.3
retrying==1.3.3
